move `chia dev` to its own subdirectory
